+++ src/controllers/deleteLine.php
@@ -2,16 +2,11 @@
 require "./database/db.php";
 try {
     if (!empty($_POST["deleteLine"])) {
-        $nombre_linea = $_POST["nombreLinea"];
-
-        $query = "SELECT ID_linea
-          FROM linea
-          WHERE nombre_linea = :nombre_linea";
-        $sql = $conn->prepare($query);
-        $sql->bindParam(":nombre_linea", $nombre_linea);
-        $sql->execute();
-        $id = $sql->fetchAll(PDO::FETCH_ASSOC);
-        
+        $id_linea = $_POST["nombreLinea"];
+        if(empty($id_linea) || !is_numeric($id_linea)) {
+            echo '<script>alert("Falata completar datos"); window.location.href = "./deleteLinePage"; </script>';
+            exit;
+        }
         foreach ($id as $key => $value) {
             $id_linea = $id[$key]["ID_linea"];
             $query = "DELETE FROM recorre WHERE ID_linea=:id_linea";
@@ -39,7 +34,7 @@ try {
             $sql->bindParam(":id_linea", $id_linea);
             $sql->execute();
         }
-        echo '<script>alert("La linea ah sido eliminada con exito"); window.location.href = "./delete"; </script>';
+        echo '<script>alert("La linea ah sido eliminada con exito"); window.location.href = "./deleteLinePage"; </script>';
     }
 } catch (Exception $error) {
     echo '<script>alert("' . $error->getMessage() . '"); </script>';
